Use a POST request on the create action instead of GET for clones
Don't clone using GET requests as they shouldn't be used for
side-effects. Instead repurpose the create action and trigger it using
JavaScript.

The reason JavaScript is used is because we can't nest multiple forms
safely across different browsers without resorting to various hacks.
Instead we keep the JavaScript-triggered form outside of the current
forms scope and provide all of the data that it requires to work on
rendering the view.