Allow addition of related content
For a single travel advice page (country), allow adding a single
related piece of content from the list of Artefacts held inside of
Panopticon.

This just tests the integration layer up to the point that an HTTP
request is made to Panopticon's API. We won't actually know about state
changes from that point; only whether the request was successful or not.