Only PUT the related Artefact IDs, not the whole Artefact
When updating related Artefacts on a given Artefact there's no need to
PUT all of the parameters of the Artefact back to Panopticon. Just the
parameter that we're changing.

Also refactored out the repetition of the WebMock.should cases to a
single should_put_artefact method to make the tests easier to read.