Use the new email signup journey…
This change points the email links on travel
advice pages to the new signup journey that uses
the email-alert-frontend.

Previously we were sending users to a generic
govdelivery page where they were prompted to check
the boxes of countries to subscribe to.

This change will require a republish of the
index and countries content items. There are rake
tasks for this.