Add GOVUK request id to payload details...
Pass the govuk request id generated at the load balancer into
the payload details as 'publishing_request_id', this will be
then be present in the downstream apps and availble to the
frontend apps for rendering in govuk:publishing-request-id
meta tags.
This end-to-end identifier will be the basis for feedback and
monitoring mechanisms to ensure content gets published and
delivered by email.