Add `PublicationCheck::Runner`
This class accepts an array of `PublishRequest` objects and 'runs' them
through each of the `DEFAULT_CHECKS` (currently ContentStoreCheck`,
EmailAlertCheck will be added in a subsequent PR)
storing the 'checked' requests into a result object.

The intention is to allow the individual checks to update the
`PublishRequest` object if they succeed. `Runner` will update the
`PublishRequest` with an indication that it has been checked allowing
them to be retried a number of times before going into 'failed' state.

`Result` will use this state to evaluate success/failure of a particular
check run and to expose a collection of failed `PublishRequests`