Add travel advice pages to the search index via Rummager
Previously, panopticon would send artifacts to Rummager. That flow is now being
deprecated in favour of a simpler approach which is to send documents directly
to rummager.

This commit adds a class that allows us to add travel advice pages to the
search index directly. It also adds a rake task `rummager:index_all` whose job
is to index travel advice pages in Rummager.

This feature requires the following in place before it can be merged:
- [ ] disable https://github.com/alphagov/panopticon/blob/master/app/models/enhancements/artefact.rb#L17

The payload now includes the following extra attributes:

- publishing_app
- rendering_app
- content_id