Run all tests for schema testing
This app currently runs only a subset of schema tests for PRs on
govuk-content-schemas. This was originally done so that the schema
tests finish quickly.

This causes two problems:

- Ordering dependencies in tests are making the schema tests fail
intermittently on CI
(https://ci.dev.publishing.service.gov.uk/job/govuk_travel_advice_publis
her_schema_tests/). This is not picked up in normal testing, because
the CI runs everything
- In other applications we've seen that people forget to tag their
tests, so that the tests aren't run on CI for schema branches. This
causes failures on master.

This commit makes the schema CI tests run the same suite as normal CI
runs. It will slow down testing, but I think it's worth the trade-off.