Publish the parent link for the index page
Parent is used to generate the breadcrumb.

Currently this is hardcoded in frontend - we will be removing this in
https://github.com/alphagov/frontend/pull/1004/ so that we can render
all frontend breadcrumbs from the content store.

Before we can do this we need to send the parent to the publishing api
again, using the rake task `publishing_api:publish` to republish the
index page.

This used to send the breadcrumb data using "passthrough hashes" but
this functionality has since been removed.
See https://github.com/alphagov/travel-advice-publisher/pull/155