Add Slimmer for preview styling.
Configure headers so that slimmer is skipped on all other routes.

We also need to enable the slimmer cache in all circumstances. Slimmer
unfortunately always uses its remote i18n support if you configure the
component templates, and makes a new request for every single string to
translate; the edit forms contain hundreds of these so the app is not
usable without the cache.