Preview change_description from previous version.
When a minor changed is published, various fields including the change
description are copied from the previous version. However now that we
are previewing drafts via the content-store we also need this field to
be populated even before publish. This change updates the presenter to
get the data from the previous version if necessary.