Conditionally draft newly created editions to the Publishing API
When a country has no previous editions a newly created edition
will have no content and will not pass schema validation.
There's no point in sending this type of draft to the Publishing API
we rarely introduce new countries so this doesn't happen often
but we should still ensure we don't attempt to send an empty draft.