Tie to mongo and logstasher versions
The db seems to hand when we update to Mongo 2.5 which I think may be
because our underlying db is a Mongo 2.4 version (but I'm not sure how
much those numbers correlate)

Logstasher installs activerecord in later versions and causes all manors
of pain. It should soon be removed from the Gemfile anyway as
govuk_app_config tracks it.