Fix tracking user needs
It looks like this hasn't been a thing with the Publishing API so
far. Checking in the Publishing API, nothing published by Travel
Advice Publisher has any associated `need_ids`.

Anyway, needs are tracked with the `meets_user_needs` link type now,
so convert the need id to the corresponding content id (by searching
in Maslow), and send that to the Publishing API.

This doesn't migrate the existing data, maybe that can be done later.