Add primary org to IndexLinksPresenter
The rake task to publish the index page will use the IndexLinksPresenter.

We are setting the Primary Publishing Organisation as a hard coded
value for every time we publish that page. This is part of work to
ensure content types that currently don't have an owning org have a
value set.