Fix edition parts re-ordering scenario
Selenium does not allow the setting of values on hidden fields
and Capybara support for sortable lists is incomplete.
See https://github.com/teamcapybara/capybara/issues/222 for details.
Use javascript to reorder parts and save the form so that we
can still test that ordering works and is rendered as expected.