Add recently published healthcheck endpoint
This commit adds a new endpoint at `/healthcheck/recently-published-editions` which returns an array of the titles and published times of travel advice published between 2 days and 1 hour ago. This endpoint will be used by email-alert-monitoring to ensure published travel advice has been sent to publishing-api and triggered an email alert.