Change recently published time range
This changes the time range used to find recently published documents so
only ones that have been published more than 90 minutes ago appear in
the list rather than 1 hour as it is currently.

We've been seeing more and more travel advice documents being published
and this had to lead to increased latency of emails. This alert fires
which means that people get called up, but there's nothing for them to
do except wait for the emails to go out. Rather than calling people, we
can give the system half an hour longer to send out the emails.