Remove default update_type for edition updates
At the moment the "significant change" radio button is pre-checked,
which probably encourages publishers to treat this as the default.

This change makes sure no default value exists for update_type.

At the moment the "significant change" radio button is pre-checked, which probably encourages publishers to treat this as the default.

We should make it unchecked by default, as per [the design system's guidance on radios](https://design-system.service.gov.uk/components/radios/).