Replace custom lint tasks and lint JS/SCSS by default
This uses the more conventional approach for GOV.UK of using a lint.rake
task that shells out to other linters.

It overrides the default task with the lint, spec and jasmine:ci tasks
(reflecting a GOV.UK convention to lint first). RSpec has a habit of
setting itself as the default task so this is cleared if this has
occurred.