Resolve Lint/ConstantDefinitionInBlock violations
This fixes the violations caused by these globally scoped classes being
defined for tests. The approach taken is to create anonymous classes and
stub constants for the test.

As part of this I fixed a couple of things in the LinkValidator tests so
that it didn't need to write to the database as part of the test. I
removed the test for validation when values changed. This wasn't part of
the validator and seemed to unit test the underlying libraries
execution.