Use the design system layout for editions
This adds in some logic that ensures users who have the `Preview Design
System` permission use the design system layout. This permission will
be used for ease of development and testing with stakeholders.

It moves existing functionality into a `legacy_edit` view. New
functionality will be built in the `edit` view.

Once happy we will then remove the permission and use this as the
default view.