Add the review endpoint
This commit adds the review endpoint and view. The review page is
just a summary list component with the parts attrs.

It also deals with adding the relevant redirects to ensure parts belonging
to published and archived editions are not editable and parts belonging
to draft editions are not reviewable (you're redirected to the edit page).