Ensure the correct params are parsed for new flow
Currently, the E2E tests are using the legacy set of view params
for the edit page. This is due the fact that the requirement for
users in non-dev and test envs to use the legacy flow established in the
get_layout method are not being extended to the
`permitted_edition_attributes` method.

I've reworked the logic around layout rendering in the controller
so it is consistent with the implementation in the countries controller
and ensured the correct params are used based on the view rendered.