Add endpoint for Reviewing a document
This commit adds the routes, controller actions and view for the review
endpoint.

It's pretty simple and just exposes a summary list component per section
of the page. It also simplifies the parts partial by pulling out
the logic related to changing the links based on the state of the
edition from the parts partial. And rendering version on the
review page that only allows you to view review page for a part.