Fix brittle unit test for LinkCheckReportCreator
The unit test for `LinkCheckReportCreator` was quite tightly coupled to the way Mongoid saved records. It meant that they started failing once Mongoid had been bumped from version 7.X to 8.X.

I've refactored these tests to cover the same behaviour as before. However rather than spying on the model to check that the 'save' method was called, the tests now check the state of the database to check that the expected records were created. In other words, we check the 'what' outcome rather than checking 'how' it was done.

This change in approach had the effect of making one of the previous tests redundant. We no longer need to check that a valid report gets saved, because the existing tests already prove that to be the case. And since it didn't cover anything that wasn't already covered, I was able to remove it without reducing the value of these tests.