test using publishing api for schemas
This demonstrates that using the branch ‘add-publishing-api-clone’ of govuk-jenkinslib, which adds support for using publishing api for schemas instead of the govuk-content-schemas repository, works. By setting the 'USE_PUBLISHING_API_FOR_SCHEMAS' parameter to true, and using the branch of publishing api which currently has schemas included ('add-content-schemas'), we can see a passing build. You can tell that the CI is using publis
hing api for schemas because we can see the message "using publishing api for content schemas" in the CI logs