[1225] - RG/MD - Move "Summary" section into parts
Summary is moving into parts to enable users to rename the section or re-order it. The publishing
UI will continue to show a summary section for editions that still have it (including historical
editions). For the rest, summary will be another part. These changes include the rake tasks that
migrate the summary section to parts.