Remove Summary from UI Views
Summary being moved into parts mean that we need to stop publishers from being able to add new summaries, however, before data is migrated, publishers should still be able to view and edit existing summaries instead of having them disappear. So we need to show summary in cases where it exists, and remove ability to edit where it doesn't have one.

Additionally historical edition needs to accurately reflect the state where summary was separately where it was, instead of as part of parts, so we still require this logic

Co-authored-by: Rylan Gooch <rylan.gooch@digital.cabinet-office.gov.uk>