Update copy for downloading map for historical editions
Travel advice publisher allows you to download & print superseeded (archived)
editions.

There's a link within this document where a user can access a more detailed
version of the map (jpg) provided in the pdf.

This links off to asset manager which handles providing them with the
resource.

FCDO would like the copy of this link updated to be more specific.