Implement worker for publishing scheduled edition
Relying on the `perform_at` to do the actual scheduling. Otherwise the responsibility of
the worker is to simply publish edition of ID passed as param.

This raises some concerns as the redis queue may not be extremely reliable and there is a
chance the jobs can get lost. We will address this risk by implementing a cronjob to run every night on scheduled editions. We can increase the frequency of the cronjob if the
redis queue proves problematic.