Remove scheduling model
We had difficulty bridging the gap between the scheduling model and the edition model as
mongo makes it difficult to implement transactions. We were creating a scenario where we
could have ended up with a `Scheduling` object but the `TravelAdviceEdition` not in a
`scheduled` state.

We decided to move the `scheduled_publication_time` back into `TravelAdviceEdition`
while maintaining the `SchedulingsController`. The controller actions are quite heavy,
especially with the custom params validation, and the scheduling concern seems
stand-alone enough to warrant the separation.

Co-authored-by: Barbara Slawinska <barbara.slawinska@digital.cabinet-office.gov.uk>