Validate presence of a change note when scheduling
We put the validation on the controller to ensure we don't get redirected to the
scheduling form if change note is missing, to match the behaviour we see on the user
flow for publishing.
Validations are also implemented on the model so we can also see the error messages if
we access the scheduling form directly and try to submit without a change note.