Fix Part validation behaviour after upgrading Mongo version
The validations for parts weren't being triggered when parent model (edition) is validated. To get around this we added an explicit custom validation method to enforce the association validation.
The issue that was happening was that only the first part validation message was being shown on the front end instead of everything whenever Parts were created in the old way outlined above.
Quick fix of tests where slugs had `_` in them which are not allowed and failing validation quietly...

Co-authored-by: Laura Ghiorghisor <laura.ghiorghisor@digital.cabinet-office.gov.uk>