Raise error when there are overdue scheduled editions
We want to publish any scheduled editions left unpublished, e.g. due to machine clocks being out of sync,
by running a cronjob every morning, which attempts to publish overdue editions by executing the worker synchronously.

If any editions are still left unpublished after running the worker, we raise a an error to ensure the team has visibility.
The intention is to observe this and gauge whether we should enable pager duty and/or increase the frequency of the cronjob.

We added a rescue inside the loop to ensure that we progress through all the editions even when one fails, and thus raise
our custom error with a message including all overdue editions. The worker can fail with a `PublishingApiNotifier::EnqueueError`
or even a state machine error. The gds-api calls themselves are inside the `PublishingApiWorker` which runs async so those are
unlikely to cause an issue.