Allow cancelling of overdue edition
If something goes wrong within the scheduled publishing worker, be it a state transition
or a downstream error coming from publishing-api, it will return without changing the edition
state (remains scheduled) and will retry.

Nonetheless, if the worker cannot fix itself, we want to allow the users to cancel the current
schedule and re-schedule.

Due to a validation restriction on the edition model, we were blocking this. The publication
time would in this case be in the past which was not allowed on the draft state. We are now allowing
it if the previous state has been scheduled.