Fix invalid date behaviour
During testing it was discovered that the incorrect format error for datetime was
showing for seemingly correct dates.

The previous implementation made use of an Integer conversion to ascertain whether the
input is a correct date/time value. Nonetheless, Integer() will interpret input
such as "0_" as octal rather than decimal, so values like "08" or "09" will not be allowed.
This was introduced during refactoring and not caught by tests.

We replaced the Integer conversion with `.to_i` and added some additional tests (valid
date input rather than covering invalid scenarios). We also brought back the digit number
regex checks, to restrict day, month, hour and minute input to 2 digits, so that a month
like "0010" is invalid.