Fix: publish model in publishing app before sending onwards when scheduling
Fix the case where there is a send_alert? check that relies on the edition being published within the publishing app.
At the moment in scheduled publishing worker, it queues a list of tasks before the edition model is published. This is in contrast to the edition controller where it is published as a model before being sent to publish on publishing API and email alerts. This means the send_alert check when sending email finds the edition to not be in published state on scheduling and hence does not send out any emails.
This commit makes the order consistent with the controller and adds a test.