Add confirmation step to email alerts trigger rake task
I have added a confirmation step, using the Thor gem, to the above mentioned rake task. This is to ensure that the person running this rake task explicitly opts in to sending out this email alert.

I have also added a stub for the prompt in the email_alert_rake_spec file