Add a confirmation step to publishing_api rake tasks
I have added a confirmation, using the Thor gem, to the following rake tasks in the publishing_api namespace:

- publish_travel_advice_index (renamed from 'publish')
- unpublish_published_edition_and_email_signup_content_item
- republish_editions
- republish_edition
- republish_email_signups:index
- republish_email_signups:editions
- republish_email_signups:country_edition

This is to ensure that the person running these tasks explicitly opts in to making these changes. I have also updated the tests to reflect the confirmation prompt, so if a user was to decline the prompt there is a test for this.